requirements.txt
fastapi[standard]==0.115.6
llama-index==0.12.6
llama-index-llms-openai==0.3.11
pytest==8.3.4
streamlit==1.41.1
mcp-use==1.3.1
fastmcp==2.6.1
python-dotenv==1.1.0
Pillow==10.4.0
requests==2.32.3